refactored code

+++ src/main/java/taptap/model/Recipient.java
@@ -1,4 +1,4 @@
 package taptap.model;
 
 public record Recipient(String firstName, String lastName, String email, String phone,
-        String nickname, String wallet, float amount) {}
+        String nickname, String wallet, double amount) {}
